Update storage settings
This changeset updates the storage settings of FEC's GovCloud RDS instances to match what is being used for the original East/West instances.  Note that these changes do not affect the staging instance at this time as we are about to resume ATO testing.  Once the testing is complete we will update the staging instance as well.

Included in this changeset are also identifier changes for the dev and prod instances (and their associated replicas) as well as enabling auto minor version upgrades.

The values used are taken from the example here: https://github.com/terraform-community-modules/tf_aws_rds/blob/master/variables.tf