Updated maintenance window and apply immediately
We recently sent some configuration updates to AWS but did not include the apply immediately parameter.  Furthermore, the maintenance window is not at an appropriate time for us.  This changeset adjusts the maintenance window across all instances and sets the apply immediately parameter to make all of our changes now.