Add custom parameter group to support config changes
This changeset adds support for creating a new custom parameter group in order to configure our PostgreSQL settings.  Specifically, we would like to enable the `max_parallel_workers_per_gather` setting in order to take advantage of parallel queries, but there might be other things we want to tweak or modify in the future.