Remove all logging configuration
Per the recommendations found in https://github.com/hashicorp/terraform/issues/9435#issuecomment-266788636, we are undoing our changes and attempting to get things setup in the correct order.  We also figured out that we do not need to attach the existing policy (we were trying to avoid that in the first place).